// frontend/src/features/earthquakes/model/usgs-geojson.types.ts
export type UsgsEarthquakeGeoJson = {
  type: 'FeatureCollection'
  metadata: {
    generated: number
    url: string
    title: string
    api: string
    count: number
    status: number
  }
  bbox: [number, number, number, number, number, number]
  features: UsgsEarthquakeFeature[]
}

export type UsgsEarthquakeFeature = {
  type: 'Feature'
  properties: {
    mag: number | null
    place: string | null
    time: number
    updated: number
    tz: number | null
    url: string
    detail: string
    felt: number | null
    cdi: number | null
    mmi: number | null
    alert: string | null
    status: string
    tsunami: number
    sig: number
    net: string
    code: string
    ids: string
    sources: string
    types: string
    nst: number | null
    dmin: number | null
    rms: number | null
    gap: number | null
    magType: string | null
    type: string
  }
  geometry: {
    type: 'Point'
    coordinates: [number, number, number]
  }
  id: string
}